\name{getDupNiCuKR}
\alias{getDupNiCuKR}
\title{Extract e0 and K-ratio data from a single duplicate Ni/Cu//PET}
\usage{
getDupNiCuKR(df, nDupl)
}
\arguments{
  \item{df}{A data frame with e0 and K-ratios for NiK and
  CuK for multiple duplicates \code{df}}

  \item{nDupl}{The duplicate to extract e0 and K-ratios
  \code{nDupl}}
}
\value{
data A data frame with e0, NiKa, and CuKa \code{data}
}
\description{
A helper function prepare data for processing with
StrataGemTwoLayersOnPET
}
\examples{
# library(rWrapStrataGem)
}
\keyword{keywords}

